requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.0
python-dotenv==1.0.1
psycopg[binary]==3.3.2
SQLAlchemy==2.0.36
alembic==1.13.2
celery==5.4.0
redis==5.0.8
itsdangerous==2.2.0
argon2-cffi==23.1.0
geographiclib==2.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
geoalchemy2==0.15.1
gunicorn==22.0.0
